Tighten ATS argument feedback scope
