+++ ethicapp-student/frontend/src/components/session-detail/phases/SemanticDifferentialPhaseView.jsx
@@ -102,7 +102,6 @@ export default function SemanticDifferentialPhaseView({
     )).sort((left, right) => left - right);
   }, [phase]);
   const previousPhaseNumbersKey = previousPhaseNumbers.join(',');
-  const isAtsBlocking = Boolean(isAtsEnabled && isAtsProcessing);
   const canOpenGroupChat = phase?.features?.chat === true
     && resolvePhaseGroupId(phase) != null;
   const canPostGroupChat = canOpenGroupChat
@@ -275,7 +274,6 @@ export default function SemanticDifferentialPhaseView({
             justification={justification}
             taskFeedback={taskFeedback}
             submitting={submittingTaskId === taskId}
-            atsBlocking={isAtsBlocking}
             previousResponsesContent={
               previousPhaseNumbers.length > 0 ? (
                 <>
